requirements.txt
google-cloud-storage==1.46.0